# http/cur_swagger.yml
openapi: "3.0.0"
info:
  title: Influx API Service
  version: 0.1.0
servers:
  - url: "http://localhost:9999/api/v2"
paths:
...
  /authorizations:
    get:
      tags:
        - Authorizations
      summary: List all authorizations
      parameters:
        - in: query
          name: userID
          schema:
            type: string
          description: filter authorizations belonging to a user id
        - in: query
          name: user
          schema:
            type: string
          description: filter authorizations belonging to a user name
      responses:
        '200':
          description: A list of authorizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorizations"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Authorizations
      summary: Create an authorization
      requestBody:
        description: authorization to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Authorization"
      responses:
        '201':
          description: authorization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /authorizations/{authID}:
    get:
      tags:
        - Authorizations
      summary: Retrieve an authorization
      parameters:
        - in: path
          name: authID
          schema:
            type: string
          required: true
          description: ID of authorization to get
      responses:
        '200':
          description: authorization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - Authorizations
      summary: update authorization to be active or inactive. requests using an inactive authorization will be rejected.
      requestBody:
        description: authorization to update to apply
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Authorization"
      parameters:
        - in: path
          name: authID
          schema:
            type: string
          required: true
          description: ID of authorization to update
      responses:
        '200':
          description: the active or inactie authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Authorizations
      summary: Delete a authorization
      parameters:
        - in: path
          name: authID
          schema:
            type: string
          required: true
          description: ID of authorization to delete
      responses:
        '204':
          description: authorization deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Permission:
      required: [action, resource]
      properties:
        id:
          type: string
        name:
          description: only exists if Permission.id field exists
          type: string
        action:
          type: string
          enum:
            - read
            - write
            - create
            - delete
        resource:
          type: string
          enum:
            - users
            - orgs
            - tasks
            - buckets
            - dashboards
            - sources
    Authorization:
      required: [permissions, orgID]
      properties:
        links:
          type: object
          readOnly: true
          example:
            self: "/api/v2/authorizations/1"
            user: "/api/v2/users/12"
          properties:
            self:
              readOnly: true
              type: string
              format: uri
            user:
              readOnly: true
              type: string
              format: uri
        id:
          readOnly: true
          type: string
        user:
          readOnly: true
          type: string
          description: name of user that created the authorization
        userID:
          readOnly: true
          type: string
          description: id of user that created the authorization
        status:
          description: if inactive the token is inactive and requests using the token will be rejected.
          default: active
          type: string
          enum:
            - active
            - inactive
        token:
          readOnly: true
          type: string
        description:
          nullable: true
          description: user defined description of the authorization
          type: string
        orgID:
          type: string
        org:
          readOnly: true
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    Authorizations:
      type: object
      properties:
        links:
          readOnly: true
          $ref: "#/components/schemas/Links"
        auths:
          type: array
          items:
            $ref: "#/components/schemas/Authorization"
...
    Link:
      type: object
      readOnly: true
      description: URI of resource.
      properties:
        href:
          type: string
          format: uri
      required: [href]
    Links:
      type: object
      properties:
        next:
          $ref: "#/components/schemas/Link"
        self:
          $ref: "#/components/schemas/Link"
        prev:
          $ref: "#/components/schemas/Link"
      required: [self]
...
    Error:
      properties:
        code:
          description: code is the machine-readable error code.
          readOnly: true
          type: string
          enum:
            - internal error
            - not found
            - conflict
            - invalid
            - empty value
        message:
          readOnly: true
          description: message is a human-readable message.
          type: string
        op:
          readOnly: true
          description: op describes the logical code operation during error. Useful for debugging.
          type: string
        err:
          readOnly: true
          description: err is a stack of errors that occurred during processing of the request. Useful for debugging.
          type: string
      required: [code, message, op, err]
...